requirements.txt
Click==7.0
decorator==4.4.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.2
six==1.12.0
validators==0.12.6
Werkzeug==0.15.4
